relocate and document functions
